+++ docs/RELEASE-NOTES.rst
@@ -11,6 +11,7 @@ Added Functionality
         * Persistence Profile support for VirtualServer, TransportServer and Policy CRs. `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
         * Added support for host in TransportServer and IngressLink CR. `Examples <https://github.com/F5Networks/k8s-bigip-ctlr/tree/master/docs/config_examples/customResource/>`_
         * Added support for multiple health monitors in EDNS resource
+    * NodePortLocal(NPL) antrea cni feature support added to Ingress and CRD Resources.
 
 Enhancements
 ````````````